autoplay: Avoid incessantly purchasing the super cheap items
If items are small cheap enough relative to our CpS, we just buy it even
if the Projected Payoff is large.  That works great on starting a new
game, but after ascending to heaven and getting lots of upgrades, it is
really annoying, because the autobuy spends all money getting hundreds
of cursors, grandmas, farms, mines, and even factories, spending money
about as fast as you can get it every second or so, instead of letting
you buy more effective items.  So, only make one of these types of
purchases every minute.