autoplay: Refactor purchase decision so cheap items won't take precedence
If we have enough to purchase the most effective item, we should buy it.
We should only purchase cheap items if we don't have enough cash for a
more effective one.

This probably only matters after re-incarnating, but it at least lets us
remove the kludgy timer we used to kina-sorta de-prioritize cheap items.