autoplay: Don't log re-purchase of buildings
Sometimes we sell or sacrifice buildings we used to have (due to Krumblor
or Pantheon or Grimoire); if so, it's annoying to see the logs filled
with re-purchase of those items.  Silence them.  But we do want to see
purchase of buildings after we ascend again, so make sure those don't
count as a "re-purchase".