autoplay: Oops!  I sure messed up that calculation, didn't I?
Comparing the number of cookies between
  a) immediately buying item and then waiting L until next lucky cookie
  b) waiting L until next lucky cookie
is comparing apples to oranges; in case (a), we have one more upgrade or
building than in case (b).  To make it fair, in (b) we should wait for
the next lucky cookie, and then buy the building.  Then we can compare
cookies to cookies.