autoplay: Base CBG vs. HofF decision on max towers
If HofF would be the better spell to cast, but we were previously selling
towers in order to be able to cast CBG more times, we shouldn't just use
CBG exclusively due to our current level of magic being low.  The fact
that we sold towers means they should be cheap and we can rebuy them, so
base the logic on the previous maximum number of towers.

When it actually comes time to cast the spell, we'll check if we have
sufficient magic at that point.