autoplay: Create a more thorough blended PP calculation
Initially, for buildings, we just used the CookieMonster PP.  Then I
realized that buildings also had value not just because of direct cookie
generation but also because they could increase the buff in building
specials, and that this effect would eventually become dominant for the
cheaper buildings.  But instead of combining the values to get an overall
value of the building, I simply looked at the value from each and took
the bigger one.  It'd be more accurate to blend the two.

Further, once you get to the point that buildings are super expensive but
you are able to cast spontaneous edifice, buildings also have value based
on how much they cost.  So, include a spontaneous edifice value and blend
it in as well.