autoplay: Avoid disabling usage strategies due to fluctuating CpS
Previously, we were turning off the usage strategies if the current cost
of a building was more than a certain percentage of (unbuffed) CpS.
However, since there are various things that can drop base CpS (e.g.
selling of cursors and wizard towers from the usage strategy, especially
due to the synergy upgrades; or Cyclius; or picking a new aura) this
meant that the usage strategy could be turned off without any actual
manual purchase of the relevant building.

To avoid this, keep track of how many buildings were automatically bought
while the usage strategy was in effect.  If the user pushes the number
of that building type over that amount, then turn the strategy off.  Then
the only way CpS will factor into things is at buying time, instead of at
both buying and comparing-to-see-if-we-should-turn-it-off time.