autoplay: Do not assume chancemakers will always be the most expensive
When working on the spontaneous edifice strategy, do not always assume that
chancemakers will be the most expensive.  They'll probably start out that
way, but after selling multiple along the way to get enough of a buffer,
eventually others catch up, and if we keep selling chancemakers, pass them
for expensiveness.  Pick the most expensive building and sell that, since it
will give us the most bang for the buck.